Initial commit
Created a screen for the game, two paddles, and a ball. The right paddle moves up and down on the up and down arrow keys. The left paddle moves up and down on the w and s keys. Currently the ball moves to the top, right corner of the screen.